LayoutTests/svg/dom/SVGLength-expected.txt

This test checks the SVGLength API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check initial length values
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_NUMBER
PASS length.value is 0
PASS length.valueInSpecifiedUnits is 0
PASS length.valueAsString is "0"

Set value to be 2px
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.valueAsString is "2px"

Check invalid arguments for 'convertToSpecifiedUnits'
PASS length.convertToSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_UNKNOWN) threw exception NotSupportedError: The operation is not supported..
PASS length.convertToSpecifiedUnits(-1) threw exception NotSupportedError: The operation is not supported..
PASS length.convertToSpecifiedUnits(11) threw exception NotSupportedError: The operation is not supported..
PASS length.convertToSpecifiedUnits('aString') threw exception NotSupportedError: The operation is not supported..
PASS length.convertToSpecifiedUnits(length) threw exception NotSupportedError: The operation is not supported..
PASS length.convertToSpecifiedUnits(svgElement) threw exception NotSupportedError: The operation is not supported..
PASS length.convertToSpecifiedUnits() threw exception TypeError: Not enough arguments.
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX

Check invalid arguments for 'newValueSpecifiedUnits'
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_UNKNOWN, 4) threw exception NotSupportedError: The operation is not supported..
PASS length.newValueSpecifiedUnits(-1, 4) threw exception NotSupportedError: The operation is not supported..
PASS length.newValueSpecifiedUnits(11, 4) threw exception NotSupportedError: The operation is not supported..
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX, 0) is undefined.
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX, 'aString') threw exception TypeError: The provided value is non-finite.
PASS length.value is 0
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX, length) threw exception TypeError: The provided value is non-finite.
PASS length.value is 0
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX, svgElement) threw exception TypeError: The provided value is non-finite.
PASS length.value is 0
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX, NaN) threw exception TypeError: The provided value is non-finite.
PASS length.value is 0
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX, Infinity) threw exception TypeError: The provided value is non-finite.
PASS length.value is 0
PASS length.newValueSpecifiedUnits(SVGLength.SVG_LENGTHTYPE_PX) threw exception TypeError: Not enough arguments.
PASS length.newValueSpecifiedUnits('aString', 4) threw exception NotSupportedError: The operation is not supported..
PASS length.newValueSpecifiedUnits(length, 4) threw exception NotSupportedError: The operation is not supported..
PASS length.newValueSpecifiedUnits(svgElement, 4) threw exception NotSupportedError: The operation is not supported..
PASS length.newValueSpecifiedUnits('aString', 'aString') threw exception TypeError: The provided value is non-finite.
PASS length.newValueSpecifiedUnits(length, length) threw exception TypeError: The provided value is non-finite.
PASS length.newValueSpecifiedUnits(svgElement, svgElement) threw exception TypeError: The provided value is non-finite.
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX

Check setting invalid 'valueAsString' arguments
PASS length.valueAsString = '10deg' threw exception SyntaxError: The string did not match the expected pattern..
PASS length.valueAsString is "2px"
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS length.valueAsString is "2px"
PASS length.valueAsString is "2px"
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS length.valueAsString = ',5 em' threw exception SyntaxError: The string did not match the expected pattern..
PASS length.valueAsString is "2px"
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS length.valueAsString = null threw exception SyntaxError: The string did not match the expected pattern..
PASS length.valueAsString is "2px"
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX

Check setting invalid 'value' arguments
PASS length.value = NaN threw exception TypeError: The provided value is non-finite.
PASS length.value = Infinity threw exception TypeError: The provided value is non-finite.
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX

Check setting invalid 'valueInSpecifiedUnits' arguments
PASS length.valueInSpecifiedUnits = NaN threw exception TypeError: The provided value is non-finite.
PASS length.valueInSpecifiedUnits = Infinity threw exception TypeError: The provided value is non-finite.
PASS length.value is 2
PASS length.valueInSpecifiedUnits is 2
PASS length.unitType is SVGLength.SVG_LENGTHTYPE_PX
PASS successfullyParsed is true

TEST COMPLETE
